requirements.txt
requests==2.32.3
pyspark==3.5.1
delta-spark==3.2.0
deltalake==0.18.2
apache-airflow==2.9.3